yt_scraper/requirements.txt
google-api-python-client
youtube-transcript-api
openai
supabase
scikit-learn
numpy
python-dotenv
tqdm 